refactor: move agent module out of retrieval and add QueryComplexity enum
- Move agent module from rust/src/retrieval/agent to rust/src/agent
- Add QueryComplexity enum to represent query complexity levels for adaptive budget selection
- Update import paths throughout codebase to reflect new agent module location
- Move BM25 scoring utilities from retrieval to utils module
- Remove redundant complexity detection module from retrieval
- Update retrieval module documentation to clarify it now provides
infrastructure rather than the main engine
