+++ rust/src/agent/command.rs
