+++ rust/src/agent/config.rs
@@ -3,6 +3,23 @@
 
 //! Configuration and output types for the retrieval agent.
 
+/// Query complexity level for adaptive budget selection.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum QueryComplexity {
+    /// Simple queries that can be solved with keyword matching.
+    Simple,
+    /// Medium complexity queries requiring semantic understanding.
+    Medium,
+    /// Complex queries requiring deep LLM reasoning.
+    Complex,
+}
+
+impl Default for QueryComplexity {
+    fn default() -> Self {
+        Self::Medium
+    }
+}
+
 use serde::{Deserialize, Serialize};
 
 /// Agent configuration.
