+++ rust/src/agent/context.rs
