+++ rust/src/agent/events.rs
