+++ rust/src/agent/mod.rs
@@ -31,7 +31,7 @@ pub mod orchestrator;
 pub mod prompts;
 pub mod subagent;
 
-pub use config::{Config, DocContext, Output, Scope, WorkspaceContext};
+pub use config::{Config, DocContext, Output, QueryComplexity, Scope, WorkspaceContext};
 pub use events::{AgentEvent, EventEmitter};
 
 /// Retrieve information from documents using the agent.
