+++ rust/src/agent/state.rs
