+++ rust/src/agent/tools/common.rs
