+++ rust/src/agent/tools/mod.rs
