+++ rust/src/client/retriever.rs
@@ -13,7 +13,7 @@ use crate::document::{DocumentTree, NavigationIndex, ReasoningIndex};
 use crate::error::{Error, Result};
 use crate::events::{EventEmitter, QueryEvent};
 use crate::llm::LlmClient;
-use crate::retrieval::agent::{self, events::EventEmitter as AgentEventEmitter};
+use crate::agent::{self, events::EventEmitter as AgentEventEmitter};
 
 /// Document retrieval client.
 ///
