+++ rust/src/retrieval/cache.rs
