--- rust/src/retrieval/cache/mod.rs
@@ -1,8 +1,0 @@
-// Copyright (c) 2026 vectorless developers
-// SPDX-License-Identifier: Apache-2.0
-
-//! Caching for retrieval operations.
-//!
-//! Reasoning cache with L1 (exact query match) hit support.
-
-mod reasoning_cache;
