+++ rust/src/retrieval/mod.rs
@@ -1,28 +1,18 @@
 // Copyright (c) 2026 vectorless developers
 // SPDX-License-Identifier: Apache-2.0
 
-//! Retrieval system for Vectorless document trees.
+//! Retrieval infrastructure — types, streaming, and caching.
 //!
-//! This module implements agent-based retrieval:
-//! - **SubAgent**: navigates a single document (ls → cd → cat → check → done)
-//! - **Orchestrator**: multi-document MapReduce (analyze → dispatch → integrate → synthesize)
+//! The actual retrieval engine lives in the top-level [`agent`](crate::agent) module.
+//! This module provides supporting infrastructure:
 //!
-//! # Architecture
-//!
-//! ```text
-//! retrieve(query, scope)
-//!     ├── Scope::Single(doc)     → SubAgent loop → Output
-//!     └── Scope::Workspace(ws)   → Orchestrator → Output
-//! ```
+//! - **Types** — `RetrieveResponse`, `SufficiencyLevel`, `ReasoningChain`, etc.
+//! - **Streaming** — `RetrieveEvent` / `RetrieveEventReceiver` for async progress
+//! - **Cache** — `ReasoningCache` for L1 query caching
 
+mod cache;
 pub mod stream;
 mod types;
 
-pub mod agent;
-pub mod cache;
-pub mod complexity;
-pub mod scoring;
-pub mod sufficiency;
-
-pub use types::*;
 pub use stream::RetrieveEventReceiver;
+pub use types::*;
