--- rust/src/retrieval/sufficiency/mod.rs
@@ -1,7 +1,0 @@
-// Copyright (c) 2026 vectorless developers
-// SPDX-License-Identifier: Apache-2.0
-
-//! Sufficiency checking types.
-//!
-//! Re-exports `SufficiencyLevel` for use by events and engine modules.
-//! The agent system has its own sufficiency logic (heuristic pre-check + LLM check).
