+++ rust/src/utils/mod.rs
@@ -5,13 +5,16 @@
 //!
 //! This module provides common utilities used across the codebase:
 //!
+//! - **BM25 scoring** — Per-field weighted text relevance scoring
 //! - **Token estimation** — Fast and accurate token counting (tiktoken-based)
 //! - **Fingerprint** — BLAKE2b content hashing for change detection
 //! - **Validation** — Pre-index source validation (file, content, bytes)
 
+pub mod bm25;
 pub mod fingerprint;
 mod token;
 pub mod validation;
 
+pub use bm25::extract_keywords;
 pub use token::estimate_tokens;
 pub use validation::{validate_bytes, validate_content, validate_file};
